Day 09
I busted back out the linked-list for this one. It really made a
difference when solving the second half.

I left solution A as an array based implementation. I would normally
merge the two identical solutions into a common function, but I wanted
to preserve the array based implementation as proof of how I approached
the problem first.